requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.1.2
python-jose==3.3.0
pgvector==0.2.4
python-dotenv==1.0.0
requests==2.31.0
aiofiles==24.1.0